Assignment 9 - Attempt #4

+++ aesd-char-driver/aesd-circular-buffer.c
@@ -147,7 +147,7 @@ void aesd_print_circular_buffer(struct aesd_circular_buffer *buffer)
 
 long long int get_aesd_buffer_length(struct aesd_circular_buffer *buffer)
 {
-	loff_t size = 0;
+	long long int size = 0;
 	int i;
 	for(i=0;i<BUFFER_SIZE;i++){
 		size = buffer->entry[i].size + size;
